refactor: Move access record handling to a new access.go file
